Change add-final-modifier action to 'refactor' kind
